Added Docker-related files for containerization.
