+++ .dockerignore
@@ -1,0 +1,3 @@
+Dockerfile
+.dockerignore
+.git
