+++ docker-compose.yml
@@ -1,0 +1,6 @@
+services:
+  web:
+    build: .
+    pull_policy: build
+    ports:
+      - "5173:5173"
